% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BCKozakVol.R
\name{BCKozakVol}
\alias{BCKozakVol}
\title{Estimate stem volume using Kozak BC equation}
\usage{
BCKozakVol(SPP, dbh, hgt)
}
\arguments{
\item{SPP}{- unique Species ("DF","RA","SS","WH","WC")}

\item{dbh}{- tree diameter at breast height (inches)}

\item{ht}{total tree height}
}
\value{
tree volume using Smalian's
#' @examples
BCKozakVol(SPP=df$SPP,dbh=df$dbh,ht=df$ht)
}
\description{
Computes estimated tree volume using Kozak Taper Equation - Kozak, A. 1988. A variable-exponent taper equation. Can. J. For. Res. 18: 1363–1368.
}
\details{
BCKozakVol - ft3 per stem
}
\author{
Premer, M.I. - Maine Forest Lab - adapated from Brian Smith - UMaine
}
