% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/braathe.index.R
\name{braathe.index}
\alias{braathe.index}
\title{Calculate Braathe Competition Index}
\usage{
braathe.index(comp.ht, focal.ht, distance)
}
\arguments{
\item{comp.ht}{- height (ft) of the competitor tree}

\item{focal.ht}{- height (ft) of the subject tree}

\item{distance}{- distance between trees (ft)}
}
\value{
Braathe index
}
\description{
Braathe
}
\examples{
braathe.index(56,57,12)
}
\references{
Braathe, P., 1980. Height increment of young single trees in relation to height and distance of neighboring trees. Mitt. Forstl. VersAnst. 130, 43–48.
}
\author{
Premer, M.I. - Maine Forest Lab
}
