% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord_convert.R
\name{coord.convert}
\alias{coord.convert}
\title{Coordinate conversion between UTM and lat/long}
\usage{
coord.convert(x, y, proj, zone)
}
\arguments{
\item{x}{x coordinate: UTM X or longitude}

\item{y}{y coodinate: UTM Y or latitude}

\item{proj}{projection of the input coordinate system: c("UTM", "longlat")}

\item{zone}{UTM zone of the input coordinate system or requested output: c(1:60)}
}
\value{
This function returns a coordinate pair in UTM of longlat projections, opposite of the input.
}
\description{
Convert between UTM coordinates to latitude and longitude or from latitude and longitude to UTM coordinates.
Coordinates will convert to UTM if "longlat" is supplied as input projection and latitude longitude if "UTM" is
supplied as the input projection.
}
\examples{
# Input coordinates in UTM zone 10, converting to long and lat
     # Seattle, WA
coord.convert(x = 550202.40, y = 5272749.67, proj = "UTM", zone = 10)

# Input coordinates in long and lat, converting to UTM zone 17
     # Orlando, FL
coord.convert(x = -81.379237, y =  28.538335, proj = "longlat", zone = 17)

}
\author{
Premer, M.I. - originally developed by Stephanie Patton
}
