% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_layers.R
\name{create_layers}
\alias{create_layers}
\title{Generate estimates of Topograhic Wetness Index, Upslope Area, and filled sinks of a Digitial Elevation Model}
\usage{
create_layers(dem, fill.sinks = TRUE, deg = 0.1)
}
\arguments{
\item{dem}{digital elevation model in .tif format}
}
\value{
Filled Sinks DEM, Upslope Area, TWI
}
\description{
Customized forestry and GIS functions
}
\examples{
create_layers(dem)
}
\author{
Premer, M.I. - Maine Forest Lab - originally a function of topmodel
}
