% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crit.break.R
\name{crit.break}
\alias{crit.break}
\title{Critical Tree Breakage Windspeed (mph)}
\usage{
crit.break(dbh, ht, thin_ratio, SPP)
}
\arguments{
\item{thin_ratio}{description}

\item{SPP}{Species Code (FVS Format) (DF - Douglas fir; WH - western hemlock; LP - loblolly pine; SP - slash pine)}

\item{DBH}{Diameter at breast height (inches)}

\item{HT}{Total Tree Height}
}
\value{
stem volume (ft3)
}
\description{
A tree and stand level estimator of windspeed critical to breakage reported in miles per hour
}
\details{
A tree and stand level estimator of windspeed critical to breakage reported in miles per hour
}
\examples{
vol.calc("BF",14.1,68)
}
\references{
Hale, S.E., Gardiner, B., Peace, A., Nicoll, B., Taylor, P., and S. Pizzirani. Comparison and validation of three versions of a forest wind risk model. Ecological Modeling & Software.68: 27-41
}
\author{
Premer, M.I. - Maine Forest Lab
}
