% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crit.throw.R
\name{crit.throw}
\alias{crit.throw}
\title{Calculates the critical windspeed for windthrow of an individual tree or stand reported in miles per hour}
\usage{
crit.throw(dbh, ht, thin_ratio, soil_depth, Species)
}
\arguments{
\item{dbh}{-Diameter at breast height (in)}

\item{ht}{- Total stem height (ft)}

\item{thin_ratio}{- Thinning ratio (ratio of growing space post/pre thinning)}

\item{soil_depth}{- Soil depth (1 for soils < 80 cm deep, 2 for soils > 80 cm deep)}

\item{Species}{- DF, Douglas fir; WH, western hemlock; LP, loblolly pine; SP, slash pine}

\item{SPP}{Species Code (FVS Format)}
}
\value{
critical speed of windthrow in mph
}
\description{
Hale, S.E., Gardiner, B., Peace, A., Nicoll, B., Taylor, P., and S. Pizzirani. Comparison and validation of three versions of a forest wind risk model. Ecological Modeling & Software.68: 27-41
}
\details{
Critical Tree Root-throw Windspeed
}
\examples{
crit.throw(dbh=10,ht=85,thin_ratio=1,soil_depth=1,"DF")
}
\author{
Premer, M.I. - Maine Forest Lab
}
