% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crown.overlap.R
\name{crown.overlap}
\alias{crown.overlap}
\title{Calculate the total area of crown overlap between two trees}
\usage{
crown.overlap(crownradi, crownradj, distance)
}
\arguments{
\item{crownradi}{- crown radius (ft) of the competitor trees}

\item{crownradj}{- crown radius (ft) of the focal/subject tree}

\item{distance}{- distance (ft) between the two stems}
}
\value{
Crown overlap (ft2)
}
\description{
The use of crown overlap is useful for calculating competition indices
}
\details{
Site Index
}
\examples{
crown.overlap(15,24,17)
}
\author{
Premer, M.I. - Maine Forest Lab
}
