% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glover.hool.index.R
\name{glover.hool.index}
\alias{glover.hool.index}
\title{Calculate Glover Hool Competition Index}
\usage{
glover.hool.index(DBH, qmd)
}
\arguments{
\item{DBH}{- DBH (inches) of the focal tree}

\item{qmd}{- Quadratic mean diameter (inches) of the plot}
}
\value{
Glover Hool index
}
\description{
Glover Hool
}
\examples{
glover.hool.index(10,13)
}
\references{
Glover GR, Hool JN. A basal area ratio predictor of loblolly pine plantation mortality. Forest Science 1979; 25(2): 275-282. 10.1093/forestscience/25.2.275
}
\author{
Premer, M.I. - Maine Forest Lab
}
