% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hegyi.index.R
\name{hegyi.index}
\alias{hegyi.index}
\title{Calculate Hegyi Competition Index}
\usage{
hegyi.index(DBH, dbhi, distance)
}
\arguments{
\item{DBH}{- DBH (inches) of the focal tree}

\item{distance}{- Distance (ft.) between trees}

\item{dbi}{- DBH (inches) of the competitor tree}
}
\value{
Heygi index
}
\description{
Hegyi
}
\examples{
hegyi.index(10,13,3)
}
\references{
Hegyi, Frank. 1974. A simulation model for managing jack pine stands. pp. 74-90. In: Growth models for tree and stand simulation, Fries, Joran. ed. Stockholm, Sweden: Royal College of Forestry.
}
\author{
Premer, M.I. - Maine Forest Lab
}
