% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCW.R
\name{MCW}
\alias{MCW}
\title{Calculate Maximum Crown Width}
\usage{
MCW(SPP, DBH)
}
\arguments{
\item{SPP}{- Species FVS Code}

\item{DBH}{- Diameter at breast height}
}
\value{
Maximum Crown Width (ft.)
}
\description{
MCW
}
\examples{
MCW("RS",8.5)
}
\references{
Russell, M. B., & Weiskittel, A. R. (2011). Maximum and Largest Crown Width Equations for 15 Tree Species in Maine.
Northern Journal of Applied Forestry, 28(2), 84–91. https://doi.org/10.1093/njaf/28.2.84
}
\author{
Premer, M.I. - Maine Forest Lab
}
