% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kozak_merchR.R
\name{merch.df}
\alias{merch.df}
\title{Virtually buck tree to logs of specified dimensions using Kozak Taper Equation,
This particular set of parameters set to Douglas-fir}
\usage{
merch.df(
  x,
  log.breaks = c(2, 5, 7, 8, 12, 51),
  log.grades = c("junk", "pulp", "cns", "ssl", "lsl", "psl"),
  display.stems = FALSE
)
}
\arguments{
\item{dbh}{- stem diameter at breast height (in)}

\item{tht}{- total stem height (ft)}

\item{cr}{- crown ratio (0-1)}

\item{hi}{- height above ground}
}
\value{
estimated Scribner volume in each designated product class
}
\description{
Kozak, Antal. (2004). My last words on taper equations. The Forestry Chronicle. 80. 507-515. 10.5558/tfc80507-4.
}
\details{
Log volume by product
}
\examples{
merch.df(14.1,58,60,40)
}
\author{
Premer, M.I. - Maine Forest Lab
}
