% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmd.R
\name{qmd}
\alias{qmd}
\title{Calculate Quadratic Mean Diameter}
\usage{
qmd(ba, tpa)
}
\arguments{
\item{ba}{basal area per acre}

\item{tpa}{trees per acre}
}
\value{
quadratic mean diameter in inches
}
\description{
This function calculates QMD (inches) given basal area and trees per acre
}
\examples{
qmd(120,310)
}
\author{
Premer, M.I. - Maine Forest Lab
}
