% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative.density.R
\name{relative.density.index}
\alias{relative.density.index}
\title{Calculate Relative Density Index}
\usage{
relative.density.index(bapa, qmd)
}
\arguments{
\item{bapa}{- basal area per acre}

\item{tpa}{- trees per acre}
}
\value{
Relative Density Index
}
\description{
RD
}
\examples{
relative.density.index(220,11.2)
}
\author{
Premer, M.I. - Maine Forest Lab
}
