% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative.spacing.R
\name{relative.spacing}
\alias{relative.spacing}
\title{Calculate Relative Spacing}
\usage{
relative.spacing(tpa, domht)
}
\arguments{
\item{tpa}{- trees per acre}

\item{domht}{- stand dominant height (ft.)}
}
\value{
Relative Spacing
}
\description{
SDI
}
\examples{
relative.spacing(310,8.5)
}
\author{
Premer, M.I. - Maine Forest Lab
}
