% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upslope.R
\name{upslope}
\alias{upslope}
\title{Calculate Upslope area a given DEM}
\usage{
upslope(dem, log = TRUE, atb = FALSE, deg = 0.12, fill.sinks = TRUE)
}
\arguments{
\item{dem}{digital elevation model in .tif format}
}
\value{
the total, relative upslope area of any given location in the raster matrix
}
\description{
Customized forestry and GIS functions
}
\details{
This function allows you to calculate the upslope area for a given raster
}
\examples{
upslope(dem)
}
\author{
Premer, M.I. - Maine Forest Lab- originally a function of topmodel
}
