% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vicary.height.R
\name{vicary.height}
\alias{vicary.height}
\title{Estimate top height of spruce/fir using Vicary, Brann, and Griffin}
\usage{
vicary.height(SPP, si, age, ba)
}
\arguments{
\item{SPP}{- Species, either "BF", or "RS"}

\item{si}{- estimated site index}

\item{age}{- estimated age of origin}

\item{ba}{- site index base age (commonly 50)}
}
\value{
Estimated mean top height
}
\description{
Vicary, Bret P.; Brann, Thomas B.; and Griffin, Raph H.. 1984. B802: Base-Age Invariant Polymorphic Site Index Curves for Even-Aged Spruce-Fir Stands in Maine. Maine Agricultural and Forest Experiment Station Bulletins 802. https://digitalcommons.library.umaine.edu/aes_bulletin/130
}
\details{
Top height
}
\examples{
height("BF",65,42,50)
}
\author{
Premer, M.I. - Maine Forest Lab
}
