% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vicary.site.R
\name{vicary.site}
\alias{vicary.site}
\title{Estimate site index of spruce/fir using Vicary, Brann, and Griffin}
\usage{
vicary.site(SPP, ht, age)
}
\arguments{
\item{SPP}{- Species, either "BF", or "RS"}

\item{ht}{- mean top height (H40)}

\item{age}{- estimated age (years)}
}
\value{
Estimated site index at base age 50
}
\description{
Vicary, Bret P.; Brann, Thomas B.; and Griffin, Raph H.. 1984. B802: Base-Age Invariant Polymorphic Site Index Curves for Even-Aged Spruce-Fir Stands in Maine. Maine Agricultural and Forest Experiment Station Bulletins 802. https://digitalcommons.library.umaine.edu/aes_bulletin/130
}
\details{
Site Index
}
\examples{
height("BF",65,42,50)
}
\author{
Premer, M.I. - Maine Forest Lab
}
