% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vol.calc.R
\name{vol.calc}
\alias{vol.calc}
\title{Calculate tree volume (ft3)}
\usage{
vol.calc(SPP, DBH, HT)
}
\arguments{
\item{SPP}{Species Code (FVS Format)}

\item{DBH}{Diameter at breast height (inches)}

\item{HT}{Total Tree Height}
}
\value{
stem volume (ft3)
}
\description{
This function allows you to calculate tree volume given species, dbh, and height
}
\examples{
vol.calc("BF",14.1,68)
}
\author{
Premer, M.I. - Maine Forest Lab
}
