% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vol_calc_NE.R
\name{vol_calc}
\alias{vol_calc}
\title{Calculate tree volume (ft3) using Honer's volume equations}
\usage{
vol_calc(SPP, DBH, HT)
}
\arguments{
\item{SPP}{Species Code (FVS Format)}

\item{DBH}{Diameter at breast height (inches)}

\item{HT}{Total Tree Height}
}
\value{
stem volume (ft3)
}
\description{
This function allows you to calculate tree volume given species, dbh, and height using Honer's volume equations. Honer, T.G. 1967. Standard volume tables and merchantable conversion factors for the commercial tree species of central and eastern Canada. Can. Dept. Forestry Rural Devel., For. Mgmt. Res. and Serv. Inst. Info. Rep. FMR-X-5.


}
\examples{
vol_calc("BF",14.1,68)
}
\author{
Premer, M.I. - Maine Forest Lab
}
