% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wykoff.ht.R
\name{wykoff.ht}
\alias{wykoff.ht}
\title{Estimate tree height using the Wykoff equation}
\usage{
wykoff.ht(SPP, DBH)
}
\arguments{
\item{SPP}{- Species code from FVS NE Varaiant}

\item{DBH}{- Diameter at breast height (in)}
}
\value{
Estimated total height
}
\description{
Wykoff, W.R.; Crookston, N.L.; Stage, A.R. 1982. User’s guide to the Stand Prognosis Model. Gen. Tech. Rep. INT-133. Ogden, UT: U. S. Department of Agriculture, Forest Service, Intermountain Forest and Range Experiment Station. 112p.
}
\details{
Tree height
}
\examples{
wykoff.ht("RS",14)
}
\author{
Premer, M.I. - Maine Forest Lab
}
